LayoutTests/imported/w3c/web-platform-tests/shadow-dom/Element-interface-attachShadow-expected.txt


PASS Check the existence of Element.attachShadow 
PASS Nodes other than Element should not have attachShadow 
PASS Element.attachShadow must throw a TypeError if mode is not "open" or "closed" 
PASS Element.attachShadow must create an instance of ShadowRoot 
FAIL Element.attachShadow must throw a NotSupportedError if the context object already hosts a shadow tree assert_throws: Calling attachShadow({mode: "open"}) twice on the same element must throw function "function () {
        var div = document.createElement('div');
        div.attachShadow({mode: "open"});
        div.attachShadow({mode: "open"});
    }" threw object "InvalidStateError: The object is in an invalid state." ("InvalidStateError") expected object "[object Object]" ("NotSupportedError")
PASS Element.attachShadow must throw a NotSupportedError for non-safelisted elements 
